API/requirements.txt
Django==2.1.7
django-cors-headers==2.5.0
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
pytz==2018.9